+++ example/test4/f_main.f90
@@ -62,13 +62,10 @@ program main
 	call f_pm_stop ("Kernel-Fast", dflop, 1)
 	call spacer (msize,n,a,b,c)
 
-	call f_pm_printprogress ("", "for checking", 0)
-	if(i.eq.3) then
-	call f_pm_reset ("Kernel-Slow")
-	endif
-	if(i.eq.4) then
-	call f_pm_resetall ()
-	endif
+	!cx call f_pm_printprogress ("", "for checking", 0)
+	!cx if(i.eq.4) then
+	!cx call f_pm_resetall ()
+	!cx endif
 
 	end do
 
@@ -76,7 +73,7 @@ program main
 
 !cx call f_pm_posttrace ()
 
-	call f_pm_print ("", 0)
+	call f_pm_print ("", "", "", 0)
 	call f_pm_printdetail ("", 0, 0)
     call f_pm_printthreads ("", 0, 0)
     call f_pm_printlegend ("")