+++ example/test4/f_openmp_do.f90
@@ -62,21 +62,13 @@ program main
 	call f_pm_stop ("Kernel-Fast", dflop, 1)
 	call spacer (msize,n,a,b,c)
 
-	call f_pm_printprogress ("", "for checking", 0)
-	if(i.eq.3) then
-	call f_pm_reset ("Kernel-Slow")
-	endif
-	if(i.eq.4) then
-	call f_pm_resetall ()
-	endif
-
 	end do
 
 	call f_pm_stop ("Loop-section", dflop*6.0, 1)
 
 !cx call f_pm_posttrace ()
 
-	call f_pm_print ("", 0)
+	call f_pm_print ("", "", "", 0)
 	call f_pm_printdetail ("", 0, 0)
     call f_pm_printthreads ("", 0, 0)
     call f_pm_printlegend ("")